Transition Shoes::Swt::Oval from module to class
Also:

- Add Shoes.configuration.backend
- Add Shoes.configuration.backend specs
- Add Shoes.configuration.backend = :swt to bin scripts

This commit contains a bunch of hacks to make backend-as-class
coexist with backend-as-module, so that we can make this transition
element by element. Hacks are mostly in Configuration, where there
are (temporarily) two ways to set the backend.

- The old way: Shoes.configuration.framework = 'swt/shoes'
- The new way: Shoes.configuration.backend = :swt

For now, we have to set both, until the transition it complete.

For specs, I created a new "mock" backend, which I envision as a
replacement for white_shoes. I think this might ultimately be the
shoes-mocks project, but for now, it's just a shim so that we
can run Shoes DSL specs without hitting a gui backend.