Fix app height calculation (was breaking specs on OS X)
Two specs were broken by commit
26f4a2588604e201dd5f6a12d7f1d17812dcd3f7, apparently only on OS X.

The spec for App#height using defaults was breaking on OS X because the
app height was including the window decorations (22px).

The spec for setting App#width was breaking on OS X because the spec set
the width to 1, but the window can't be resized smaller than 69px. This
changeset uses a higher width for the spec.

In addition, this fix lets us remove some hacks around background
rendering.