Rework how we track text styles for fragments
Instead of an Array[[Range, Text]], use a Hash{Range => [Text]}. Now,
on the DSL level, the fragment styles are organized by the substring
they apply to, so we have lots less processing to do in the backend.
Hooray for no more "nested_styles"! See spec/shoes/text_block_spec.rb
for an example of the expected format.