Fix crash with line offsets
Under certain circumstances, we end up trying to iterate to the last text
offset that a TextLayout reports to us. Unfortunately that last element was
indexed past the end of the layout's line metrics, so we'd crash. Should only
be including offsets up to, but not including, the last one for consideration.