Potential fix for relative vs absolute path
Following instructions on pirate_game against Shoes4 master, found that the
bin/shoes script wouldn't run right. In particular, where we do a
`cd ./$NEXT_DIR` would fail when running against the aliased command pointing
to the gem installed binary.

I'm highly suspicious that there is either a better way to do this on OS X, or
that this won't be right when we finally get a gem-installable version of
shoes, but figured I'd put this out there while I was looking at it.