Support passing click:"/url" option to Link
Passing an option value :click with a string is treated a "url" that we hand
off to the App#visit method for dispatching.

This addresses part of #588 in which clicking the links wasn't bringing up the
incident text it should have. There are other layout issues there, so it won't
be fully resolved by this change.