Simple menu fixes for compatibility
There are a couple of post-alpha issues that were preventing this sample from
working, and also a compatibility change in how we reference the Shoes classes
for use from `style`. This commit should get the sample running on master, and
it can be revised when later functionality allows.

Also removes the copied simple-menu1.rb which was presumably a Shoes4
compatible version of the original simple-menu.rb