Trigger redrawing on game of life
The existing game of life code wasn't properly triggering element redraws when
the style of a cell changed states. This meant that clicking run would go, but
only show you the results when you forced a redraw by other means (i.e.
resizing the window).

This commit adds a redrawing trigger for the common `style` call and updates
the sample to be more cautious about style-setting to avoid masses of updates
all at once.