Safer in_bounds? check
Original version of the bounds checking was, for whatever strange reason
made sense to me at the time, modifying the bounds we got back to do a
contains? check. This is odd and potentially hazardous (if layouts
didn't return a clean copy of bounds).

Additionally in reviewing it, I realized the logic fit much more cleanly
within the FittedTextLayout class itself, so I moved it along with
removing the strangeness.