Centralize more font/styling in FittedTextLayout
Since the FittedTextLayout class was already responsible for some of the
resource management because of creating Swt::TextLayout objects, I'm
moving toward having it be the central point that holds onto the SWT
resources for text block rendering.

This required some test updates, as things were a bit too tightly
coupled. There are more changes on the way (for instance the fitter
shouldn't need to talk to the text block to get a new FittedTextLayout
anymore), but those will wait for a later commit.