Clear transform after painting each object
The transform which is set by the Shoes::Swt::Common::Painter, if it's
provided by the object, was being allowed to persist across to
additional painting operations on that graphics context.

Most art objects didn't actually have a transform set, but line does so
elements following it would get mixed up by having the line's transform
applied to them.