Move Shoes::SHOES_ICON -> Shoes::ICON
This was never used in the DSL, and it may often require
backend-specific code (i.e. an app packaged in a jar needs different
file-loading code from one simply run from the command line).

Let's package the icon file with shoes-dsl, and set a default in the dsl
layer, so there's something reliable to start from. But then in the
backend, set our own icon constant as appropriate, and use that for
backend work.