Beginnings of text centering
This commit will return a special text segment type for a centered
TextBlock. This segment consumes as much space as possible to provide
the default "centering" experience most users would expect for text.

Work remains for dealing with situations when this isn't the first
element on a given line (for now we'll be bumping down), but this gets
the basic centering in place.