Cleanup for link click registration weirdness
For whatever reason when I last revised the Link class it ended up with
this weird Proc wrapper being added at the SWT layer to make sure links
got updated properly.

It's much simpler to just make sure the DSL passes along changes to the
block we want called on click.

Because of a bit of special link behavior this doesn't actually use the
Shoes::Common::Clickable module, but it does leverage
Shoes::Swt::Common::Clickable on the backend, and is closer to the
standard click implementation.