Trigger redraw on hover/leave for slot
To get a proper redraw after the hover/leave on a slot, we needed to
pull the proc execution into the mouse_hovered/mouse_left methods so we
could hook those with the RedrawingAspect.

This moves the specs for this to the DSL level since the `Shoes::Slot`
now takes care of executing the procs that it's holding as well.