Be more specific about what to include from bin
When we started including the executables in shoes.gemspec, the
existence of shoes.bat (which is a symlink on *nix style systems)
started causing problems when we try to build the gem.

As it turns out, we don't actually need that file in bin as part of the
gem's packaging--it is copied from ext/install during installation
anyway, and only linked into bin as a convenience for running from
source on Windows.

With this change, we can package again on *nix systems and hopefully
continue to work properly with the changed paths on Windows.