Fix uninstall to properly clean up after us
We generate two non-Ruby files in the Gem bin dir--shoes (during gem
install) and shoes-backend (during the first run or picking a shoes
backend).

The shoes binary wasn't getting uninstalled because we weren't putting
the lib/rubygems_plugin.rb file into the gem itself.

shoes-backend was a recently added file that got missed on adding to
cleanup when I created it.