Fix setting top slot dimensions
We were drawing dimensions for the top slot from resize events, but
these ended up forcing a scrolling top slot to the size of the
containing window, not its contents, which was wrong.

Responsibility for this is now moved to the `ShoesLayout` itself, which
does a better job of it.

Fixes #617.