Get samples/expert-funnies.rb running again
With the background issue resolved, there's no reason for this sample to
be exclued.

Except it doesn't work :( XKCD appears to have discontinued the RSS
element that we were drawing our image from, and the code didn't handle
it gracefully.

So now we look to see if we get that parsed out right before adding a
section, so things are a bit more resilient.