Draw our own background for text
Turns out that SWT's TextLayout only draws the background around the
characters that it draws, even when the layout is explicitly set to a
greater width.

When we're filling in the first of multiple text segments, we actually
want the full space filled even though the text probably doesn't eat it
all up. Given that, we have to perform the fill ourselves more directly.