Load system colors during backend initialization
This fixes the defaulting problems we were having and prevents us from
having to do the color loading multiple times (which could happen in the
prior scheme if you had multiple windows.)

Downside, it does remove the defaulting in the DSL layer since there's
not an elegant way to get the value when we need it. Assigning it in the
backend just ends up cleaner.