Restore spec in Core suite, fix type failure
Fixes #1206

It was noted that the spec failed but our suite didn't. That was because
our `Dir` for files to run was overly constrainted and missed it in the
'shoes-core/spec' directory.

After restoring that to run, we still have the exception. This was just
a change at some point in behavior from `ArgumentError` to `LoadError`. I
think `LoadError` is actually more representative given what's happening
there, so I modified the test to match.