Memoize and handle resets for translation styles
Because we're using the style system for transmitting the translate
value (which greatly simplifies things), it does mean that the
individual elements expose translate as a value that can be accessed or
modified. In the case that someone does modify it, we clear our
memoization of the translation values so they get recalculated on the
next pass.