Pick up gems from Bundler
If we detect the presence of Bundler when packaging, we'll try to use
its list of gems (minus 'shoes' which causes us problems and isn't
necessary to install). This allows Rubyists to use the typical means of
declaring their dependencies in a Ruby app, but works nicely even
without a Gemfile around too.